Moved decoration flags to DecorationEnums.h
