+++ src/Engine/Objects/DecorationEnums.h
@@ -2,6 +2,8 @@
 
 #include <cstdint>
 
+#include "Utility/Flags.h"
+
 /**
  * Enum of all decorations in the game, e.g. trees, trash heaps, pedestals, etc.
  *
@@ -238,3 +240,36 @@ enum class DecorationId : int32_t {
     DECORATION_TREE45 = 226,               // tree45, tree
     DECORATION_TREE46 = 227,               // tree46, tree
 };
+using enum DecorationId;
+
+
+enum class DecorationDescFlag : uint16_t {
+    DECORATION_DESC_MOVE_THROUGH = 0x0001,
+    DECORATION_DESC_DONT_DRAW = 0x0002,
+    DECORATION_DESC_FLICKER_SLOW = 0x0004,
+    DECORATION_DESC_FICKER_AVERAGE = 0x0008,
+    DECORATION_DESC_FICKER_FAST = 0x0010,
+    DECORATION_DESC_MARKER = 0x0020,
+    DECORATION_DESC_SLOW_LOOP = 0x0040,
+    DECORATION_DESC_EMITS_FIRE = 0x0080,
+    DECORATION_DESC_SOUND_ON_DAWN = 0x0100,
+    DECORATION_DESC_SOUND_ON_DUSK = 0x0200,
+    DECORATION_DESC_EMITS_SMOKE = 0x0400,
+};
+using enum DecorationDescFlag;
+MM_DECLARE_FLAGS(DecorationDescFlags, DecorationDescFlag)
+MM_DECLARE_OPERATORS_FOR_FLAGS(DecorationDescFlags)
+
+
+enum class LevelDecorationFlag : uint16_t {
+    LEVEL_DECORATION_TRIGGERED_BY_TOUCH = 0x01,
+    LEVEL_DECORATION_TRIGGERED_BY_MONSTER = 0x02,
+    LEVEL_DECORATION_TRIGGERED_BY_OBJECT = 0x04,
+    LEVEL_DECORATION_VISIBLE_ON_MAP = 0x08,
+    LEVEL_DECORATION_CHEST = 0x10,
+    LEVEL_DECORATION_INVISIBLE = 0x20,
+    LEVEL_DECORATION_OBELISK_CHEST = 0x40,
+};
+using enum LevelDecorationFlag;
+MM_DECLARE_FLAGS(LevelDecorationFlags, LevelDecorationFlag)
+MM_DECLARE_OPERATORS_FOR_FLAGS(LevelDecorationFlags)
