+++ src/Engine/Objects/DecorationList.h
@@ -8,25 +8,9 @@
 
 #include "Library/Color/Color.h"
 
-#include "Utility/Flags.h"
 #include "Utility/Memory/Blob.h"
 
-enum class DecorationDescFlag : uint16_t {
-    DECORATION_DESC_MOVE_THROUGH = 0x0001,
-    DECORATION_DESC_DONT_DRAW = 0x0002,
-    DECORATION_DESC_FLICKER_SLOW = 0x0004,
-    DECORATION_DESC_FICKER_AVERAGE = 0x0008,
-    DECORATION_DESC_FICKER_FAST = 0x0010,
-    DECORATION_DESC_MARKER = 0x0020,
-    DECORATION_DESC_SLOW_LOOP = 0x0040,
-    DECORATION_DESC_EMITS_FIRE = 0x0080,
-    DECORATION_DESC_SOUND_ON_DAWN = 0x0100,
-    DECORATION_DESC_SOUND_ON_DUSK = 0x0200,
-    DECORATION_DESC_EMITS_SMOKE = 0x0400,
-};
-using enum DecorationDescFlag;
-MM_DECLARE_FLAGS(DecorationDescFlags, DecorationDescFlag)
-MM_DECLARE_OPERATORS_FOR_FLAGS(DecorationDescFlags)
+#include "DecorationEnums.h"
 
 struct DecorationDesc {
     std::string name; // Internal name (e.g. "dec03").
